Add auth cookie to websocket requests
